Replace old ptree JSON library
